+++ lib/features/scan/scan_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:vigil_ai/features/link_scanner/link_scanner_screen.dart';
 import 'package:vigil_ai/features/message_scanner/message_scanner_screen.dart';
 import 'package:vigil_ai/features/qr_scanner/qr_scanner_screen.dart';
+import 'package:vigil_ai/features/upi_checker/upi_checker_screen.dart';
 
 class ScanScreen extends StatelessWidget {
   const ScanScreen({super.key});
@@ -70,9 +71,10 @@ class ScanScreen extends StatelessWidget {
               title: const Text("UPI Fraud Check"),
               trailing: const Icon(Icons.arrow_forward_ios),
               onTap: () {
-                ScaffoldMessenger.of(
+                Navigator.push(
                   context,
-                ).showSnackBar(const SnackBar(content: Text("Coming Soon")));
+                  MaterialPageRoute(builder: (_) => const UpiCheckerScreen()),
+                );
               },
             ),
           ),